selectField test
