+++ tests/SubmitField.test.jsx
@@ -1,33 +1,18 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import {SelectField} from '../src/components/SelectField';
+import {SubmitField} from '../src/components/SubmitField';
 
-describe('SelectField', () => {
+describe('SubmitField', () => {
     it('should receive props and call onChange method on change value', () => {
         const mockFunction = jest.fn();
-        const testError = ['testError'];
-        const errorStyles = {
-            className: 'errorClassName'
-        };
-        const options = ['option1', 'option2', 'option3'];
         const props = {
-            name: 'firstName',
-            label: 'first name',
-            onChange: mockFunction,
-            error: true,
-            errors: testError,
-            errorsStyles: errorStyles,
-            className: 'testComponent'
+            submit: mockFunction,
+            value: 'Submit'
         };
-        const wrapper = mount(<SelectField {...props} />);
-        expect(wrapper.props().name).toBe(props.name);
-        expect(wrapper.props().label).toBe(props.label);
-        expect(wrapper.props().onChange).toBe(props.onChange);
-        expect(wrapper.props().error).toBe(props.error);
-        expect(wrapper.props().errors).toBe(props.errors);
-        wrapper.setProps({options});
-        expect(wrapper.props().options).toBe(options);
-        wrapper.find('select').simulate('change', 'option3');
+        const wrapper = mount(<SubmitField {...props} />);
+        expect(wrapper.props().value).toBe(props.value);
+        expect(wrapper.props().submit).toBe(props.submit);
+        wrapper.find('button').simulate('click');
         expect(mockFunction.mock.calls.length).toBe(1);
     });
 });
